1. Introduced properties file ( use file for now, may move over to db in future) 2. Genome files are now loaded as resources 3. Basic ChIPSeq intersection is working, for now only returning first 100 records. I'll add table controls on the next commit 4. If any chromsome naming/range errors are encountered when running the intersection, they can be viewed by clicking the exclaimation point at \     the top of the reuslts table. 5. Query dropdowns are resticted by user, but not by other checkboxes, will fix this in next commit.
git-svn-id: svn://svn.code.sf.net/p/biominerqt/code-0/trunk@95 86439291-5e3c-406c-9808-073398cf6135

+++ src/main/java/hci/biominer/model/QueryResult.java
@@ -11,8 +11,8 @@ public class QueryResult {
   private String sampleConditions;
   private String analysisSummary;
   private String coordinates;
-  private BigDecimal log2Ratio;
-  private BigDecimal FDR; 
+  private Float log2Ratio;
+  private Float FDR; 
 	
 	public QueryResult() {
 	}
@@ -57,19 +57,19 @@ public class QueryResult {
     this.coordinates = coordinates;
   }
 
-  public BigDecimal getLog2Ratio() {
+  public Float getLog2Ratio() {
     return log2Ratio;
   }
 
-  public void setLog2Ratio(BigDecimal log2Ratio) {
+  public void setLog2Ratio(Float log2Ratio) {
     this.log2Ratio = log2Ratio;
   }
 
-  public BigDecimal getFDR() {
+  public Float getFDR() {
     return FDR;
   }
 
-  public void setFDR(BigDecimal fDR) {
+  public void setFDR(Float fDR) {
     FDR = fDR;
   }
